// src/comum/enums/status-ordem-servico.enum.ts
export enum StatusOrdemServico {
  ABERTA = 'aberta',
  AGUARDANDO_COLETA = 'aguardando_coleta',
  COLETADA = 'coletada',
  EM_HIGIENIZACAO = 'em_higienizacao',
  EM_MANUTENCAO = 'em_manutencao',
  CONTROLE_QUALIDADE = 'controle_qualidade',
  PRONTA_PARA_ENTREGA = 'pronta_para_entrega',
  ENTREGUE = 'entregue',
  CANCELADA = 'cancelada',
}